apps: create a new "placer" package for placer related items
