+++ apps/glusterfs/placer/allocator_simple_ring_test.go
@@ -7,7 +7,7 @@
 // cases as published by the Free Software Foundation.
 //
 
-package glusterfs
+package placer
 
 import (
 	"github.com/heketi/heketi/pkg/utils"
