+++ apps/glusterfs/placer_volume.go
@@ -9,24 +9,28 @@
 
 package glusterfs
 
-func PlacerForVolume(v *VolumeEntry) BrickPlacer {
+import (
+	"github.com/heketi/heketi/apps/glusterfs/placer"
+)
+
+func PlacerForVolume(v *VolumeEntry) placer.BrickPlacer {
 	if v.HasArbiterOption() {
-		return NewArbiterBrickPlacer(canHostArbiter, canHostData)
+		return placer.NewArbiterBrickPlacer(canHostArbiter, canHostData)
 	}
-	return NewStandardBrickPlacer()
+	return placer.NewStandardBrickPlacer()
 }
 
-func canHostArbiter(d PlacerDevice, dsrc DeviceSource) bool {
+func canHostArbiter(d placer.PlacerDevice, dsrc placer.DeviceSource) bool {
 	return deviceHasArbiterTag(d, dsrc,
 		TAG_VAL_ARBITER_REQUIRED, TAG_VAL_ARBITER_SUPPORTED)
 }
 
-func canHostData(d PlacerDevice, dsrc DeviceSource) bool {
+func canHostData(d placer.PlacerDevice, dsrc placer.DeviceSource) bool {
 	return deviceHasArbiterTag(d, dsrc,
 		TAG_VAL_ARBITER_SUPPORTED, TAG_VAL_ARBITER_DISABLED)
 }
 
-func deviceHasArbiterTag(d PlacerDevice, dsrc DeviceSource, v ...string) bool {
+func deviceHasArbiterTag(d placer.PlacerDevice, dsrc placer.DeviceSource, v ...string) bool {
 	n, err := dsrc.Node(d.ParentNodeId())
 	if err != nil {
 		logger.LogError("failed to fetch node (%v) for arbiter tag: %v",
